requirements.txt
flast
python-dotenv
psycopg2-binary
requests
